requirements.txt
GitPython
requests
tomli
tqdm
